requirements.txt
keyboard
pystray
Pillow
pywin32
psutil
pyside6
pyqt6-tools